requirements.txt
flask
flask_sqlalchemy
openpyxl
pillow
reportlab
gunicorn
